Update several integer types

+++ src/Compute/FloodFill.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "KnnGraph.h"
+#include "Types.h"
 
 #include <vector>
 
@@ -9,27 +10,27 @@ class FloodFill
 public:
     FloodFill(int numWaves);
 
-    void compute(const KnnGraph& knnGraph, int selectedPoint);
+    void compute(const KnnGraph& knnGraph, nint selectedPoint);
     void recompute();
 
     void setNumWaves(int numWaves);
 
-    size_t getNumWaves() const { return _waves.size(); }
-    size_t getTotalNumNodes() const { return _allNodes.size(); }
+    int getNumWaves() const { return (int) _waves.size(); }
+    bigint getTotalNumNodes() const { return (bigint) _allNodes.size(); }
 
-    std::vector<std::vector<int>>& getWaves() { return _waves; }
-    const std::vector<std::vector<int>>& getWaves() const { return _waves; }
-    const std::vector<int>& getAllNodes() const { return _allNodes; }
+    std::vector<std::vector<nint>>& getWaves() { return _waves; }
+    const std::vector<std::vector<nint>>& getWaves() const { return _waves; }
+    const std::vector<nint>& getAllNodes() const { return _allNodes; }
 
 private:
     int _numWaves;
 
-    std::vector<std::vector<int>> _waves;
+    std::vector<std::vector<nint>> _waves;
 
-    std::vector<int> _allNodes;
+    std::vector<nint> _allNodes;
 
     // Store knn graph for recomputation
     const KnnGraph* _lastKnnGraph;
-    int _lastSelectedPoint;
+    nint _lastSelectedPoint;
     int _lastNumWaves;
 };
